# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface OkResult {
  ok: boolean;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface CodeInput {
  code: string;
}

export interface PasswordInput {
  /** @minLength 6 */
  password: string;
}

export interface KeyCodeInput {
  code: string;
}

export interface KeyCustomerInput {
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerPhone?: string | null;
}

export interface ManagerInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export type PromotionInputPlan = typeof PromotionInputPlan[keyof typeof PromotionInputPlan];


export const PromotionInputPlan = {
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
  lifetime: 'lifetime',
} as const;

export interface PromotionInput {
  plan: PromotionInputPlan;
  /** @minimum 50 */
  priceCents: number;
  /**
     * @minimum 1
     * @maximum 8760
     */
  durationHours: number;
  /** @nullable */
  bannerText?: string | null;
}

export type PromotionPlan = typeof PromotionPlan[keyof typeof PromotionPlan];


export const PromotionPlan = {
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
  lifetime: 'lifetime',
} as const;

export interface Promotion {
  id: string;
  plan: PromotionPlan;
  priceCents: number;
  /** @nullable */
  bannerText: string | null;
  endsAt: string;
  createdAt: string;
  active: boolean;
}

export interface PlanPriceInput {
  /** @minimum 50 */
  priceCents: number;
}

export interface ValidateInput {
  code: string;
  fingerprint: string;
}

export type KeyGenerateInputPlan = typeof KeyGenerateInputPlan[keyof typeof KeyGenerateInputPlan];


export const KeyGenerateInputPlan = {
  trial: 'trial',
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
  lifetime: 'lifetime',
} as const;

export interface KeyGenerateInput {
  plan: KeyGenerateInputPlan;
  /**
     * @minimum 1
     * @maximum 500
     */
  quantity: number;
  /** @nullable */
  userEmail?: string | null;
}

export type AuthResultStatus = typeof AuthResultStatus[keyof typeof AuthResultStatus];


export const AuthResultStatus = {
  authenticated: 'authenticated',
  code_required: 'code_required',
  pending: 'pending',
  rejected: 'rejected',
} as const;

export type SessionUserRole = typeof SessionUserRole[keyof typeof SessionUserRole];


export const SessionUserRole = {
  admin: 'admin',
  manager: 'manager',
  client: 'client',
} as const;

export type SessionUserStatus = typeof SessionUserStatus[keyof typeof SessionUserStatus];


export const SessionUserStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface SessionUser {
  id: string;
  name: string;
  email: string;
  role: SessionUserRole;
  status: SessionUserStatus;
}

export interface AuthResult {
  status: AuthResultStatus;
  user?: SessionUser | null;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  manager: 'manager',
  client: 'client',
} as const;

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  keyCount: number;
  createdAt: string;
}

export type KeyPlan = typeof KeyPlan[keyof typeof KeyPlan];


export const KeyPlan = {
  trial: 'trial',
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
  lifetime: 'lifetime',
} as const;

export type KeyStatus = typeof KeyStatus[keyof typeof KeyStatus];


export const KeyStatus = {
  inactive: 'inactive',
  active: 'active',
  expired: 'expired',
  revoked: 'revoked',
} as const;

export interface Key {
  id: string;
  code: string;
  plan: KeyPlan;
  status: KeyStatus;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  userEmail?: string | null;
  /** @nullable */
  deviceFingerprint?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  activatedAt?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
}

export interface PlanCounts {
  trial: number;
  daily: number;
  weekly: number;
  monthly: number;
  lifetime: number;
}

export interface AdminStats {
  totalUsers: number;
  pendingUsers: number;
  approvedUsers: number;
  totalKeys: number;
  activeKeys: number;
  inactiveKeys: number;
  expiredKeys: number;
  revokedKeys: number;
  planCounts: PlanCounts;
}

export interface ResetResult {
  ok: boolean;
  plan: string;
}

export interface ValidateResult {
  valid: boolean;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  plan?: string | null;
  /** @nullable */
  expiresAt?: string | null;
}

export type GetUsersParams = {
status?: GetUsersStatus;
};

export type GetUsersStatus = typeof GetUsersStatus[keyof typeof GetUsersStatus];


export const GetUsersStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export type GetKeysParams = {
plan?: GetKeysPlan;
status?: GetKeysStatus;
};

export type GetKeysPlan = typeof GetKeysPlan[keyof typeof GetKeysPlan];


export const GetKeysPlan = {
  trial: 'trial',
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
  lifetime: 'lifetime',
} as const;

export type GetKeysStatus = typeof GetKeysStatus[keyof typeof GetKeysStatus];


export const GetKeysStatus = {
  inactive: 'inactive',
  active: 'active',
  expired: 'expired',
  revoked: 'revoked',
} as const;