# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me/keys:
    get:
      operationId: getMyKeys
      tags: [me]
      summary: List keys owned by the current client
      responses:
        "200":
          description: Keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Key"
        "401":
          $ref: "#/components/responses/Error"
  /me/keys/activate:
    post:
      operationId: activateKey
      tags: [me]
      summary: Claim and activate a key by code (starts the timer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeyCodeInput"
      responses:
        "200":
          description: Activated key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Key"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /me/keys/{id}/reset-device:
    post:
      operationId: resetMyKeyDevice
      tags: [me]
      summary: Reset the device binding of one of my keys
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Updated key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Key"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /me/keys/{id}/customer:
    patch:
      operationId: updateMyKeyCustomer
      tags: [me]
      summary: Set or clear the customer label (name/email) of one of my keys
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeyCustomerInput"
      responses:
        "200":
          description: Updated key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Key"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /me/trial:
    post:
      operationId: generateTrial
      tags: [me]
      summary: Generate a free 15-minute trial key (requires owning at least one paid key)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeyCustomerInput"
      responses:
        "201":
          description: Trial key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Key"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    KeyCodeInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
    KeyCustomerInput:
      type: object
      properties:
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
...
    Key:
      type: object
      required: [id, code, plan, status, createdAt]
      properties:
        id:
          type: string
        code:
          type: string
        plan:
          type: string
          enum: [trial, daily, weekly, monthly, lifetime]
        status:
          type: string
          enum: [inactive, active, expired, revoked]
        userId:
          type: ["string", "null"]
        userEmail:
          type: ["string", "null"]
        deviceFingerprint:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        activatedAt:
          type: ["string", "null"]
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...