# lib/api-zod/src/generated/types/keyCustomerInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface KeyCustomerInput {
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerPhone?: string | null;
}